toolsdb: include failsafe against removing admin accounts
In fixing the problem with stale credentials in maintain-dbusers,
there was introduced a small risk of removing an admin account during
processing.  This just adds one more safeguard against that.

Bug: T188680

+++ modules/role/files/labs/db/maintain-dbusers.py
@@ -60,10 +60,12 @@ import os
 import time
 import pymysql
 import random
+import re
 from hashlib import sha1
 import subprocess
 import netifaces
 
+#  '^[spu][0-9]'
 PROJECT = 'tools'
 PASSWORD_LENGTH = 16
 PASSWORD_CHARS = string.ascii_letters + string.digits
@@ -380,7 +382,7 @@ def create_accounts(config):
     Find hosts with accounts in absent state, and creates them.
     """
     acct_db = get_accounts_db_conn(config)
-
+    username_re = re.compile('^[spu][0-9]')
     for host in config['labsdbs']['hosts']:
         labsdb = pymysql.connect(
             host,
@@ -406,7 +408,8 @@ def create_accounts(config):
                     except pymysql.err.InternalError as err:
                         # When on a "legacy" server, it is possible there is an old
                         # account that will need cleanup before we create it anew.
-                        if err.args[0] == 1396 and grant_type == 'legacy':
+                        if err.args[0] == 1396 and grant_type == 'legacy' and \
+                                username_re.match(row['mysql_username']):
                             labsdb_cur.execute(
                                 "DROP USER '{username}'@'%';".format(
                                     username=row['mysql_username']
